Adds better type defs for navigators/screens. Adds POC for DemoScreen demos.
