+++ boilerplate/app/screens/DemoComponentsScreen/demos/index.ts
@@ -1,0 +1,1 @@
+export * from "./TextDemo"
