+++ boilerplate/app/screens/index.ts
@@ -1,5 +1,5 @@
 export * from "./WelcomeScreen"
-export * from "./demo/components-screen"
-export * from "./demo/demo-screen"
+export * from "./DemoDebugScreen"
+export * from "./DemoComponentsScreen/DemoComponentsScreen"
 export * from "./error/error-boundary"
 // export other screens here
