+++ boilerplate/ignite/templates/navigator/NAME-navigator.tsx.ejs
@@ -12,7 +12,7 @@ const Stack = createStackNavigator<<%= props.pascalCaseName %>NavigatorParamList
 export const <%= props.pascalCaseName %>Navigator = () => {
   return (
     <Stack.Navigator screenOptions={{ cardStyle: { backgroundColor: "transparent" }, headerShown: false, }}>
-      <Stack.Screen name="demo" component={DemoScreen} />
+      <Stack.Screen name="Demo" component={DemoScreen} />
     </Stack.Navigator>
   )
 }
